+++ Light.h
@@ -7,7 +7,7 @@ public:
 	Light();
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat Xdir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity);
 
-	void useLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+	void useLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
 
 	~Light();
